pyproject.toml
[tool.poetry]
name = "bookmarks-service"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.12.3"
celery = "^5.4.0"
django = "^5.0.6"
djangorestframework = "^3.15.1"
drf-spectacular = "^0.27.2"
psycopg = {extras = ["binary"], version = "^3.1.19"}
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.9"

[tool.ruff]
line-length = 120
src = ["src"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
